requirements.txt
torch>=2.2
torchvision>=0.17
transformers>=4.43
accelerate>=0.33
einops>=0.7
tqdm>=4.66
pillow>=10.3
numpy>=1.26
scikit-image>=0.24
opencv-python>=4.10
matplotlib>=3.9
sentencepiece==0.2.1
lpips
diffusers